LayoutTests/imported/w3c/web-platform-tests/intersection-observer/unclipped-root-expected.txt


PASS Test that border bounding box is used to calculate intersection with a non-scrolling root. 
FAIL First rAF. assert_equals: entries.length expected 1 but got 0
FAIL target.style.transform = 'translateY(195px)' assert_equals: entries.length expected 2 but got 0
